+++ tests/test_concurrent.py
@@ -8,14 +8,13 @@ from asphalt.core import Context
 from asphalt.core.concurrent import executor
 
 
-@pytest_asyncio.fixture
-async def context():
-    async with Context() as ctx:
-        yield ctx
+@pytest.fixture
+def context() -> Context:
+    return Context()
 
 
 @pytest_asyncio.fixture
-async def special_executor(context):
+async def special_executor(context: Context) -> ThreadPoolExecutor:
     executor = ThreadPoolExecutor(1)
     context.add_resource(executor, "special", types=[Executor])
     context.add_teardown_callback(executor.shutdown)
@@ -31,8 +30,9 @@ async def test_executor_special(context, use_resource_name, special_executor):
     def check_thread(ctx):
         assert current_thread() is executor_thread
 
-    executor_thread = special_executor.submit(current_thread).result()
-    await check_thread(context)
+    async with context:
+        executor_thread = special_executor.submit(current_thread).result()
+        await check_thread(context)
 
 
 @pytest.mark.asyncio
@@ -41,8 +41,9 @@ async def test_executor_default(event_loop, context):
     def check_thread(ctx):
         assert current_thread() is not event_loop_thread
 
-    event_loop_thread = current_thread()
-    await check_thread(context)
+    async with context:
+        event_loop_thread = current_thread()
+        await check_thread(context)
 
 
 @pytest.mark.asyncio
@@ -58,10 +59,11 @@ async def test_executor_worker_thread(event_loop, context, special_executor):
         assert current_thread() is not special_executor_thread
         return runs_in_special_worker(ctx, current_thread())
 
-    event_loop_thread = current_thread()
-    special_executor_thread = special_executor.submit(current_thread).result()
-    retval = await runs_in_default_worker(context)
-    assert retval == "foo"
+    async with context:
+        event_loop_thread = current_thread()
+        special_executor_thread = special_executor.submit(current_thread).result()
+        retval = await runs_in_default_worker(context)
+        assert retval == "foo"
 
 
 @pytest.mark.asyncio
@@ -70,8 +72,9 @@ async def test_executor_missing_context(event_loop, context):
     def runs_in_default_worker():
         pass
 
-    with pytest.raises(RuntimeError) as exc:
-        await runs_in_default_worker()
+    async with context:
+        with pytest.raises(RuntimeError) as exc:
+            await runs_in_default_worker()
 
     exc.match(
         "the callable needs to be called with a Context as the first or second positional "